# Cache resolved suppliers for expressions