The idea of the PR is that usually users use the same or similar expressions in cycle. In this case their suppliers could be cached to save time on resolving.
It leads to some improvement like
before
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  1829.098 ± 16.737  ops/ms
JmhTest._letterifyExpression        thrpt    5  1775.457 ±  9.980  ops/ms
JmhTest._optionsExpression          thrpt    5  1052.901 ± 16.139  ops/ms
JmhTest._regexifyExpression         thrpt    5   366.460 ±  3.624  ops/ms
JmhTest.firstName                   thrpt    5   994.682 ±  8.504  ops/ms
JmhTest.initFaker                   thrpt    5   250.549 ±  2.140  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5   873.172 ± 20.675  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   233.320 ±  1.172  ops/ms
```

after
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  2822.148 ± 15.956  ops/ms
JmhTest._letterifyExpression        thrpt    5  2882.468 ± 45.041  ops/ms
JmhTest._optionsExpression          thrpt    5  1632.520 ± 31.800  ops/ms
JmhTest._regexifyExpression         thrpt    5   406.823 ±  0.950  ops/ms
JmhTest.firstName                   thrpt    5  2323.425 ± 13.366  ops/ms
JmhTest.initFaker                   thrpt    5   237.218 ± 10.073  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  1355.382 ±  4.804  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   464.029 ±  3.581  ops/ms
```